Correcting

+++ malloc_free/0-create_array.c
@@ -17,7 +17,7 @@ char *create_array(unsigned int size, char c)
 	if (size < 1)
 		return (NULL);
 
-	for (; i < size; i++)
+	for (; i < (int) size; i++)
 		ptr[i] = c;
